+++ crates/forge/tests/e2e/completions.rs
@@ -26,7 +26,7 @@ fn test_generate_completions_unsupported_shell() {
         std::env::set_var("SHELL", "/bin/unsupported");
     }
 
-    let snapbox = Command::new(snforge_test_bin_path()).arg("completion");
+    let snapbox = Command::new(snforge_test_bin_path()).arg("completions");
 
     let output = snapbox.assert().failure();
 
@@ -38,26 +38,4 @@ fn test_generate_completions_unsupported_shell() {
             "
         ),
     );
-}
-
-#[test]
-fn test_deprecated_alias() {
-    for variant in Shell::value_variants() {
-        let shell = variant.to_string();
-
-        let output = Command::new(snforge_test_bin_path())
-            .arg("completion")
-            .arg(shell.as_str())
-            .assert()
-            .success();
-
-        assert_stdout_contains(
-            output,
-            indoc!(
-                r"
-                # [WARNING] Command `snforge completion` is deprecated and will be removed in the future. Please use `snforge completions` instead.
-                "
-            ),
-        );
-    }
 }