+++ crates/sncast/tests/e2e/completions.rs
@@ -39,22 +39,3 @@ fn test_generate_completions_unsupported_shell() {
         ),
     );
 }
-
-#[test]
-fn test_deprecated_alias() {
-    for variant in Shell::value_variants() {
-        let shell = variant.to_string();
-        let args = vec!["completion", shell.as_str()];
-
-        let output = runner(&args).assert().success();
-
-        assert_stdout_contains(
-            output,
-            indoc!(
-                r"
-                # [WARNING] Command `sncast completion` is deprecated and will be removed in the future. Please use `sncast completions` instead.
-                "
-            ),
-        );
-    }
-}